test(net.redis): :white_check_mark: use xunit to unit test
