--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/DataAnnotationUtil.cs
@@ -1,23 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace CodeDesignPlus.Net.Redis.Test.Helpers;
-
-public static class DataAnnotationsExtensions
-{
-    public static IList<ValidationResult> Validate<T>(this T data)
-    {
-        if (data == null)
-            throw new ArgumentNullException(nameof(data));
-
-        var results = new List<ValidationResult>();
-
-        var validationContext = new ValidationContext(data, null, null);
-
-        Validator.TryValidateObject(data, validationContext, results, true);
-
-        if (data is IValidatableObject @object)
-            @object.Validate(validationContext);
-
-        return results;
-    }
-}
