--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Models/AppSettings.cs
@@ -1,8 +1,0 @@
-﻿using CodeDesignPlus.Net.Redis.Options;
-
-namespace CodeDesignPlus.Net.Redis.Test.Helpers.Models;
-
-public class AppSettings
-{
-    public RedisOptions? Redis { get; set; }
-}
