--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/ca.crt
@@ -1,32 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFlTCCA30CFDMsLt43W7pMtp5gG3dCInjXHDoIMA0GCSqGSIb3DQEBCwUAMIGG
-MRcwFQYDVQQKDA5Db2RlRGVzaWduUGx1czEeMBwGA1UEAwwVQ2VydGlmaWNhdGUg
-QXV0aG9yaXR5MQswCQYDVQQGEwJDTzESMBAGA1UECAwJQm9nb3TDg8KhMRIwEAYD
-VQQHDAlCb2dvdMODwqExFjAUBgNVBAsMDUlUIERlcGFydG1lbnQwHhcNMjMxMDE5
-MjMxODA1WhcNMzMxMDE2MjMxODA1WjCBhjEXMBUGA1UECgwOQ29kZURlc2lnblBs
-dXMxHjAcBgNVBAMMFUNlcnRpZmljYXRlIEF1dGhvcml0eTELMAkGA1UEBhMCQ08x
-EjAQBgNVBAgMCUJvZ290w4PCoTESMBAGA1UEBwwJQm9nb3TDg8KhMRYwFAYDVQQL
-DA1JVCBEZXBhcnRtZW50MIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEA
-3eqsaVLZ4K0o1QgFnfgwSZ5xdKK5IaBLoQrK1B0V/ZjX/MnGTr86/7KiU1DvnBQ1
-TA9trqs+MOUIHA/0ltOR+ThPJ5H2nsjXyn8VUVA4lcqM7yEhiKN5y6Pk0m2M+nMU
-DuLTrN8eAy11GlnbZlMR/8BEUVWjTMnpv7QeEl6w/o+Qtq0DPNlv/cvsei948qC7
-B56/owgP+wggOc9deqrUOmFMQBPgKC0IuGeGdRLDuIgdPjb3+RIppU80iz6m3LNn
-DhHctew8u/bfuIEnFWSvxh1myDNxzKtyzF5docQ6PrDPHLOmq86F8wCwDfccFbCe
-tRIhjaPtbojMDrusPJgvjWJqyVP1UtmN+K08oUUjxbzk6epMxsC/NWNJEj2w3bKd
-MqFYUJ5wR3OrRgvsGDKJQl44nJNZhAzKRSwgEdThd82esWJlnfJD9B10w+mJeH0a
-o1rJ1vUdvbbefUHXePBMgN8utCDqUSSJ7oONhVkSbmbadLBj2W4ekmffHgXqs5u2
-8vIz2H23RfWs1Rq/I2IwZB02nuWT2HI6JRTsQs1xvK2DqCR1V3lQfe/95D3yJfOT
-K3HRNHza9yzl1a9pJAID3/IJD+/dQptWv5vcfQeTCSM0L+XdXW0sWjcJbk9XsPwc
-zn3Rv6oJWF11QIihXTgNiVzamix7OuxBrbKMqJ9mwi8CAwEAATANBgkqhkiG9w0B
-AQsFAAOCAgEAudQZVrrw+4tUsgcdGtnOwjFzR4ggMPy3E+c8Mf8/qJ4ja2c14chf
-0Ne1AS+q7SuQarPk5/4EkqxZFtAOE2uQmJ3S2sjbyD6T4vEALn5/iZlIReMQ7ANe
-RlQ3tM4v+INy6Zi2Pr8hemQJtWa1YbD2Rzy5xoMNFbXVuqG+1CWf58JWWGpCXoxp
-HcnUsU8jtVreUHAr0Qo/kSkv5ztTYhD/k2P2+o/WaImRBX7ClUkqJOTENQFGr4se
-RIsKRSsUrCD4DbX95a64eXGO6kVtnmVxpeMk+MJYTKDYRGHf/ziA+aZeczAL1nN3
-7y6F71YIqJh2GUCKxjr0zh9Fv8chgNl66BxcyjolZggXXv2r9z/Fp9CB9TGm5579
-hzhnFmT/SGGPTmhHwOfLoZCxc0qCt1WPR06Ri4gVmezEjwZL+Swt9RcBWb2oLoNG
-cxY3YEeabY02irSVgCiqstb/QcO+pKgI9FbVwIUN0Qj4kKUVAfxd4Iwac/DvyxLA
-WsjoSgQoWyaq0CypfRPHzjEXc9hH7kYbbVRolDUfHsbjLvyTNe2u1hWI4TCwLA7G
-V/amo2ojuZ9fLJOAB0rvE7qzci8lMmyuazxStN7+xPeJIF3KB68oUyQmE2JuLm+p
-NxDMAbAbFwvCbT2O2D75eqYv4TfayHvAo8Vji51oXfNnFNy15vGEdVc=
------END CERTIFICATE-----
