--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/ca.key
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJRAIBADANBgkqhkiG9w0BAQEFAASCCS4wggkqAgEAAoICAQDd6qxpUtngrSjV
-CAWd+DBJnnF0orkhoEuhCsrUHRX9mNf8ycZOvzr/sqJTUO+cFDVMD22uqz4w5Qgc
-D/SW05H5OE8nkfaeyNfKfxVRUDiVyozvISGIo3nLo+TSbYz6cxQO4tOs3x4DLXUa
-WdtmUxH/wERRVaNMyem/tB4SXrD+j5C2rQM82W/9y+x6L3jyoLsHnr+jCA/7CCA5
-z116qtQ6YUxAE+AoLQi4Z4Z1EsO4iB0+Nvf5EimlTzSLPqbcs2cOEdy17Dy79t+4
-gScVZK/GHWbIM3HMq3LMXl2hxDo+sM8cs6arzoXzALAN9xwVsJ61EiGNo+1uiMwO
-u6w8mC+NYmrJU/VS2Y34rTyhRSPFvOTp6kzGwL81Y0kSPbDdsp0yoVhQnnBHc6tG
-C+wYMolCXjick1mEDMpFLCAR1OF3zZ6xYmWd8kP0HXTD6Yl4fRqjWsnW9R29tt59
-Qdd48EyA3y60IOpRJInug42FWRJuZtp0sGPZbh6SZ98eBeqzm7by8jPYfbdF9azV
-Gr8jYjBkHTae5ZPYcjolFOxCzXG8rYOoJHVXeVB97/3kPfIl85MrcdE0fNr3LOXV
-r2kkAgPf8gkP791Cm1a/m9x9B5MJIzQv5d1dbSxaNwluT1ew/BzOfdG/qglYXXVA
-iKFdOA2JXNqaLHs67EGtsoyon2bCLwIDAQABAoICAB9V9rrin5sfExp4c66ZmrFH
-5cvApIqv2K4TWOO4fbVIbWCn2fr0UYTvY7DcchE54J1uZPWv/uv8Pfe/IFxh4LAZ
-OELP/3aZdAXcib/ofNet+SuqeSfoXuzspjA5QZnQln7+9EE/ZOdcS7ezrjbrOP9C
-Vi3z/UdSZabpF1akYggixp5SR5hCjef/xn3UZpbq2ljE8ljiQmYXVibuS0wPLkEA
-zI8UdL0c6tPQjocekLOQc4H9qhNUx2ZX4lqyjJuyhjVRkWHA9iyGkOhmmFm9f308
-A++g0AtvaZpGeEs86wqU1edpeM9e2MGhEbskbyX4QYOJCA2OSHB+HEO3qCZgSP/B
-r0YAETENssGT7nQt6kjPTq+UA48RDSXJKbNK0FVRPeZOfQqASIg2RQcHJy571cpK
-V0JHeAutp6jLsz4ZcDdRMB4V1NIYKOhEvHJ2PIBwQkv3WSBW3And2twxSzW5l5xN
-Q41DMktUkS/f42kyHvsvXbp1b/5HPsWFEfR8L23b14uK4XufihKqsrexoBCbOEjj
-AkniFMIMsObbFIqDnuwCMlLxK5XFQfh8ba2ugkVhg5kJROQSuqLLlIEVAx973lYO
-zhtcZScdA5h7NrXP3xNx4m5bCSW3VuMG86de6cYGm9l1iRiadSe5Ttu+OlvQBfC/
-VOII1oU4VOj/Bgu9RODxAoIBAQD63Pi9J1tzGfL5NbsZRGyGBTbbKkr5K9S6h3o8
-Ek4IlnX/noIGxblegBdJNdcnd2H0uXtqnTQIjq5TqR9Xu6UPUp0BSlDGj1xSx0JP
-T6FG41xDGa+mP3C6X6dZFj7pyaaO8rRSzZvu4uufxPaNDfT4x1gCLV4eKomrJmI4
-+aZ+fCDZHpXxm65ZuAdBZITcnThHZbXGa2j2JQnBm8uk0zu8+nFhf3Yy162lYKrc
-o344YQBfB3Q79Hrmw83ZzUttpO2px3matj2IMKWsai8sFr7BJTVl3TM5tR8iIGgg
-fugC1eNiudEChg2yiRrodKK47PhltJuGs8xQmD8PleWTxESfAoIBAQDidfbIVbhl
-SsFEuk0cHVlwDvS8O8UjrwSDf7XgqhaomH0F9V2+xsflI4L/0sodL69TH/qEO17I
-fBZfTHd4W58AgG3mHlw73+IcDWePyefmOdEFnLu/0OfNRKBSQEwQDg1qW2tX9zDx
-FrB15bGt62vjcV68VhJcVyO1ZpmYgPVq8WkLW1XbCOQLXSS9Ks+Pj0lFHTj8Agfd
-Q8E9nYmEmLRj0DB0Ws04HNctgzQu91NSqELZV8O0F+jjOVr9VQOzwCSmAJVT1VuX
-bIcB6Am+YLAQnLo3WjOe6CWBFQ4Y8VoHcH6Lc2ILnBKo7ttHRYSq0HNiy9SGMzAa
-PZGfhySMsIhxAoIBAQCwThp2fbGUvDj6ElwTDvaiXgnGj/uvELHqAgrTKx4yhOQb
-KWBCuDVmB+nVpwQ2FX3/i1gEfsUx+mzkGSkQ9Ky1jx8O6BbK5bhB6Q5eRR2Yf3O5
-KNU8prtoEI/BWZQPpIvtTP+UbjJX+p0UZU3NpnBzY04kAZRy6PxmD4zadAIiRMSl
-Mwy4h+xJ278loyqJDemerAnLkHNtAmzP59mAVoBWeUCz/+unIdNFNiH+/NxZbTrg
-jpGPDWxOwG/s0Y5gWZg6SRnatJecHmxhNyM9/iJEBg14MLSuqu9dgZ8Mr4QJch7j
-pKoS9vmrbxY9JrOx+cE9YWgojDZcshMntfZqGncLAoIBAQDCSztkPkdqgGy5Lq4h
-xPyLLAnFxfIp32sb92JwUv4FwG7jwdSgbaPrcEBghNVrnRHrW7Bv6xAzKmN3d/+4
-f8M3hvjDSqEPrdItB0MnFq+j008GZaaft/hpIMv8RbX8wFG6Qp0iZ67R6gmT3XJI
-KYfzJ72EEMTH/OtiI5j8QwhMrLrXR5Gtzz74kP2IX0adgO34Ik5VyNg+enc9CgCG
-hw2d5WztJ0qTLgtfvQX0lbYjYMGZ/7akuS7SZM8wMtRuUBghZLa5qoBetnpOpGpX
-sFTeq9kIzJclaz9iNFJYDBqV72pi8qfYFhvcXLI9v+NHOnkE9qRsT1ah7NkX6bbu
-DjfBAoIBAQCKI1ihVA84A2h7+O9AWRDcGs31CGuaeJl/D4HcTe6YgFxxlUZsdtdP
-OT4cA+Gw+3446sYyUPv3YMBdKQn2xFvL/K00xa+wf2VOzdA/volhBE46pFSD3kD7
-ZoKxCjT7H7OaDrU+omKEaquuRkPZsn5daNM507EX+qpdFNCIChYDn048gPYeV3T5
-/kCqh+uBIgCvQzOFoBglADuRFZGJ4Ba1WNMNB9/tDjdgJqqrfcvEEb25o6wmfNmO
-Kh+2RzjYYImgZFK0YJ/hug/4NfU1kSwiFr+Uidll0DfE0a3180sju6SWhPwwa87Z
-rCA39FXwDTiho9YYd4lTCcaFJorsBsw+
------END PRIVATE KEY-----
