--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/ca.txt
@@ -1,1 +1,0 @@
-621A7BD9E31968BDDB3C00312F60C9FC9375A525
