--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/client.crt
@@ -1,32 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFiDCCA3CgAwIBAgIUYhp72eMZaL3bPAAxL2DJ/JN1pSQwDQYJKoZIhvcNAQEL
-BQAwgYYxFzAVBgNVBAoMDkNvZGVEZXNpZ25QbHVzMR4wHAYDVQQDDBVDZXJ0aWZp
-Y2F0ZSBBdXRob3JpdHkxCzAJBgNVBAYTAkNPMRIwEAYDVQQIDAlCb2dvdMODwqEx
-EjAQBgNVBAcMCUJvZ290w4PCoTEWMBQGA1UECwwNSVQgRGVwYXJ0bWVudDAeFw0y
-MzEwMTkyMzE4MDZaFw0yNDEwMTgyMzE4MDZaMH0xFzAVBgNVBAoMDkNvZGVEZXNp
-Z25QbHVzMRUwEwYDVQQDDAxyZWRpcy1jbGllbnQxCzAJBgNVBAYTAkNPMRIwEAYD
-VQQIDAlCb2dvdMODwqExEjAQBgNVBAcMCUJvZ290w4PCoTEWMBQGA1UECwwNSVQg
-RGVwYXJ0bWVudDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJudpjMh
-UslH03SdzvFfqqrkUKaHHPbV+RwDapZOSJ959R59651lFcyuJk3Y/PQFgbd7bsYI
-9+7qMSBcdnd8utScgPuhzRCJEo7NXKKl0hWzI8axeVvYvRv/8BozgY34ekosDPej
-qT4MEXSldG4swzfwe/HFp8REYliR9OwE/IBNxjfH+Y5qPF7WDIFVjjchgFKBeCIv
-Eoy5MhdulB/2fAlTgI3SgcEk/Mz6Vx8kkcYlLzoNq2johXR6mqL7KmDkra7krDNZ
-5NqxC/YbRrkN+bRb3yuSxsUP5NpqYqJxaAnEvDA9mgqXGgA4HkSBRqoeeDp2ttOx
-8BnnrN34jdYBT00CAwEAAaOB9TCB8jALBgNVHQ8EBAMCBaAwEQYJYIZIAYb4QgEB
-BAQDAgeAMB0GA1UdDgQWBBRpf3i89cdMqK761HXRQsurkwy8JDCBsAYDVR0jBIGo
-MIGloYGMpIGJMIGGMRcwFQYDVQQKDA5Db2RlRGVzaWduUGx1czEeMBwGA1UEAwwV
-Q2VydGlmaWNhdGUgQXV0aG9yaXR5MQswCQYDVQQGEwJDTzESMBAGA1UECAwJQm9n
-b3TDg8KhMRIwEAYDVQQHDAlCb2dvdMODwqExFjAUBgNVBAsMDUlUIERlcGFydG1l
-bnSCFDMsLt43W7pMtp5gG3dCInjXHDoIMA0GCSqGSIb3DQEBCwUAA4ICAQBHUEEx
-V+Wm4MEm/VoLxr5fdW0/OxRvQNthT1x5FUwVkddbvgeukT9hn9kPaNj+XLcxq437
-Ej9t+7URWJVM7HApLBbPFdjUOjcCs5JuP7dGJMZmU5aVgIFUd0Hca8+a0uYBCn4j
-W13WxrfDT5LIM71dIG29lFNmwY2jFKniQs13HR5GFg08HG+81p35/f2YZ9tvSoZ2
-ucRH4IqnfwiMTMJ+sldTc8wrmKdbw5d2VSaa11l1TIXIged2RM7lLrV+hQBkKk3F
-MHCbXXfK4hN7tSmIwLSSbBnykEXcMc++M815EusfXGAInh43L+Rx9H8zGY18yelX
-gtvKzBuImXDVy2uU3CvyUREysq549iF7u0FeaISdzqu2KwNptG9cfMIl3FyChEMk
-EOsdQ5KUZKROHSNCOeUbnED5sqxg5NpDLL0hwm9btcrdfFR/dYz+d9DyCGKbarX0
-m1SQ4QT9qZZs9Z9WxU+KxoMwlE404IlA8418u7qF399CMxwOSLcoE+QEO7+8jIsO
-KLg6fgHzJZuPZdgeyEw8WcLSNYs5dyA0805TcAsX7Fm1asc6lvRKcl7O6TrSIeMc
-Bs9JZi0vGGDRnxicQlD/pXRG7x4ugfZV8ty4LMN+JdwJpuDjv+yfvYMFqFjCzHYF
-haizrAzqDfCcYDCtzhkJ4Gtbwjmq5TY+sDmF3g==
------END CERTIFICATE-----
