--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/client.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCbnaYzIVLJR9N0
-nc7xX6qq5FCmhxz21fkcA2qWTkifefUefeudZRXMriZN2Pz0BYG3e27GCPfu6jEg
-XHZ3fLrUnID7oc0QiRKOzVyipdIVsyPGsXlb2L0b//AaM4GN+HpKLAz3o6k+DBF0
-pXRuLMM38HvxxafERGJYkfTsBPyATcY3x/mOajxe1gyBVY43IYBSgXgiLxKMuTIX
-bpQf9nwJU4CN0oHBJPzM+lcfJJHGJS86Dato6IV0epqi+ypg5K2u5KwzWeTasQv2
-G0a5Dfm0W98rksbFD+TaamKicWgJxLwwPZoKlxoAOB5EgUaqHng6drbTsfAZ56zd
-+I3WAU9NAgMBAAECggEABqdWzWCIi87/OsEVRvB1DtSEoZzXqggxHA0GAteQYbKT
-E/2Eqa4CS3D2IKWn/ykNyFIbROq/lclmq1W4wlgkxuV1hbd/2falF/0Dj81w7k5C
-t9gBHTe8XOpJvQe5wN/AdCkZBCz3wLg/C31C6AO69HUwX4COMU/5N9nvHnkCG4bm
-stZcOWHhGGsGauBqSwI0MraIMZ7y/rd0SjOstcObYFexwIeiNuTbteGlCjisrdpa
-4HELuP7qgePqHJhIkJV56QdJ+4J18m7CKKKo/YFZ/YtNz8QCJaBHIwI3RR/bORnO
-CsU1FDMnya13ggEWsIaMosoIEqohbVPI94UMEdp/8QKBgQDRqNMytsjbY5nWz/wR
-k5QJbXimd4oi0xGcqHEo3vsC4wChzddYxPM2asR4ULQGPlhz6Tm7VXpK9zDjsI86
-fMwhQyYjEjoWup04mFsbeXZvqcLeh2C12QkSpqoRaGuB6UI8SGr4ACn5oGZszOPQ
-QDl6rKW9CfuYLiFTd7LCknrXyQKBgQC+At5jcCCwJXJemYFVd/HZDRLoR2a92Akn
-QcVhAxqk1q8C2KSXZ2GLqg+4avobz//yf87iBKnih4raRWokGVLGQ+Tf4Hp/P10t
-fDhHIL4fgy2rkMxoch3N1iNarnpB0uIrMdh2k/1lm73pm/tgP7nqaf0dO2tE8+6U
-qbVs1tlFZQKBgFQ6qAaSymThQde8HP+Os4WKFlOCaNuOwl+WiEesNVm2ivV2NdKy
-5Sbyg9X2TizH9U8z3n61W3ksj9aZgyEzK+sdCABE/VNOS4WGJ+wEwyEfKWuvJZpJ
-BE0a6e4ubip15KdFTLwvjK191EhXlBqBAtaW78rfbNV5iuf2ibmDkZiJAoGAPzXh
-GZ9xoGe0tnfGqNIwKI7K3W7zZPB7iw28Rdusf23T5tR/XMbNkpLOjAE6lwoJEXkn
-L9VxVqTUsmwpNT0fZkmr2DZrSe2gZ8NNTYw882rGRK708UIVXCaCuGYW0dL/3nLE
-mh39TFIVSaPUal6I0MzVoWXk95lWoyoZ6s1GlmkCgYB0ctpfavvHFvSOVfypYeLb
-bZlaxYXpikhdwVdhQgh4YbEg/Yu537Qw/1QrreZh2EXBUETFgLRCwwghac/cFtOH
-n5FQo9jAZUL/VM2uoQieB6kWLR0aXTx3zxUdy52cw55cnBR86KShlfePIMyJzpjl
-24WZ8UORoZpCnF/7Fx2B9w==
------END PRIVATE KEY-----
