--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/openssl.cnf
@@ -1,6 +1,0 @@
-[ server_cert ]
-keyUsage = digitalSignature, keyEncipherment
-nsCertType = server
-[ client_cert ]
-keyUsage = digitalSignature, keyEncipherment
-nsCertType = client
