--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/redis.crt
@@ -1,27 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIEhDCCAmwCFGIae9njGWi92zwAMS9gyfyTdaUlMA0GCSqGSIb3DQEBCwUAMIGG
-MRcwFQYDVQQKDA5Db2RlRGVzaWduUGx1czEeMBwGA1UEAwwVQ2VydGlmaWNhdGUg
-QXV0aG9yaXR5MQswCQYDVQQGEwJDTzESMBAGA1UECAwJQm9nb3TDg8KhMRIwEAYD
-VQQHDAlCb2dvdMODwqExFjAUBgNVBAsMDUlUIERlcGFydG1lbnQwHhcNMjMxMDE5
-MjMxODA2WhcNMjQxMDE4MjMxODA2WjB2MRcwFQYDVQQKDA5Db2RlRGVzaWduUGx1
-czEOMAwGA1UEAwwFcmVkaXMxCzAJBgNVBAYTAkNPMRIwEAYDVQQIDAlCb2dvdMOD
-wqExEjAQBgNVBAcMCUJvZ290w4PCoTEWMBQGA1UECwwNSVQgRGVwYXJ0bWVudDCC
-ASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMxENg1gjJUbtaWwor/tDC9Y
-zU99vGeAddXvceZj7EMVKHui13Jk4BUsGnEwLc+984axeoWTk4ZXarKD3XcNlDzN
-oZlM7mPBEoqm1v9DxScW6Zqu9ImtTxOrrNyIG50mtwWun88AsAztad8Z4BcWuLUA
-IJUOLRo+0x6uuaYcHe2IJyoYTxKJhOAICBqMjEolPDhhXCHS533R8Fr0rcFRh7KF
-EY4rAzYKAz1V6bcpkzukZk8HZ74IVdpkcyhAxOfa1ujMWa1XLC2S1GqDSBfXiztH
-GbX6argR5k5OUzqDH/XGXvxHrnCt8romLRBsSXnhlYeKVwifoaMMgcoRsvSLXXcC
-AwEAATANBgkqhkiG9w0BAQsFAAOCAgEAc+u+3HNyuMdZzC6CdkujkUiTdr16ee5p
-AJA8tIcERLdBhDlVQ1YvPInNohG9LoihGPTfBuX4R/S1VHPvmFaMd6Ry0pS8FYIj
-C5Zru6A3ScRnOelDi+VPr25slUXlbpnNVkeb96jV2Kl26BGWJViIVMCrVA3T75QK
-GYuxDvHIGJKBBXzAKf1fnnkR1NrC3bvGWqkEA9XxVRDrNHFjuqgijEUyQ4Mtu7fP
-n66W8yHzG3KbbX86/44i2wjLDG9iRw9bgyjxn0B1Ual+Q8anAmw2a3Va7AMKoO3C
-3gaZA/UnEdSzxzWMx9WdfGsYbsfHMckeUM/C8Xay76xhZe1pglMZ8vC1PcZVLhKw
-4hIamsLgQUL//wdAFtAxpxD8RS6ExtZZzdHt2ZTvIFfG0ohYGh/KDpIPcl0YeEs1
-F6S3sy8tsW+B+sGWmknkyRWYTmJkVhVLEq9ZOI69ZIaDKSzV+NnRCkrphYbrg+Rj
-unZArUsvW8kUUIupJt2+G7oAKPxKyOVSlNQQuUmM9rrnn7+AqaUXYhGqCfWJyaqE
-05hFyhg3bEjhC6Q1D0Xjphq4eV+RVSbBKLy+JKxyQ0IcyjzFfrEjZrJwjmDHwSLK
-l8+GF4zCcG6F8nxAY8RvmTKTUnSuzodnebVzBR/X1CJ9V2a3Wf99VxAlYjoyZ28p
-M73M8tSxsNI=
------END CERTIFICATE-----
