--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/redis.dh
@@ -1,8 +1,0 @@
------BEGIN DH PARAMETERS-----
-MIIBDAKCAQEA4IlFh+VdJ5c8zRxkegO+99rum7fBxWYcpkDZi6bCn9K34I0CWKPm
-A2pI+qPMUI++lm7bvAwvjdcLQAOTUOia8xUXsa3hwL13XsThMtbba9Sd8gg9tPPq
-A0MyNLcaglHCgckzMZNdcWSf8CJAVR7Wbcq+BzB7II2ZHzb3aAkXhtcL0/cnVMpH
-qw+ZOb59v0kR089x/q0FFaIFcPkTxFBY20NLccoCVdQB8oTOLvM124fvTaam8Dq7
-Lio/MWBsK3gXiJgKsmtx3dRaUPlxrSjyTp6KSzv3Nrwq31m6rZ6SzDpO3x71Odu7
-TOleMGZw8wV1WAPpZJH6yEje23+0ppforwIBAgICAOE=
------END DH PARAMETERS-----
