--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/redis.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDMRDYNYIyVG7Wl
-sKK/7QwvWM1PfbxngHXV73HmY+xDFSh7otdyZOAVLBpxMC3PvfOGsXqFk5OGV2qy
-g913DZQ8zaGZTO5jwRKKptb/Q8UnFumarvSJrU8Tq6zciBudJrcFrp/PALAM7Wnf
-GeAXFri1ACCVDi0aPtMerrmmHB3tiCcqGE8SiYTgCAgajIxKJTw4YVwh0ud90fBa
-9K3BUYeyhRGOKwM2CgM9Vem3KZM7pGZPB2e+CFXaZHMoQMTn2tbozFmtVywtktRq
-g0gX14s7Rxm1+mq4EeZOTlM6gx/1xl78R65wrfK6Ji0QbEl54ZWHilcIn6GjDIHK
-EbL0i113AgMBAAECggEASv+OcwnshADJCH1avCoua6HovBN33e6yFSsVwFgW0qiN
-lRt8jEFSbyOiz8U3RMvDAsTeLfY517z37XRyF3iyYULE3xL6gkDM7ICGdvHsjMYi
-Xkt3eKdb3eaOEOiebuxcjacIgQoJGa3/zczj1MZgagCSdfffAoBnpp/ThRRUvrWy
-8AUDXXRy8uVv8LI1t7evNLDDk2F7Z0F4TaRPw3IluDLgwhrygMP8zfI97XYKYTw4
-BfoBoCsbne0OBv8hZtIx/U43qrgileHjOQ0BLnfoXz24pHRzCLeZK7x36kGEgzSB
-GZA/Y657EULk8T9tNa331ysBAOjvYB4nU813zO3EhQKBgQD3Qp637ZxVlmz51KfT
-z7w/g3Tv0gsvHYOsVn2ZqSAn46P4SZeZko1v5Qrj5MIy+uo0AOPNZyYryvBcxqmi
-8Ykw6EKSBgEDaIoMjTwpQ2h9anmZmd6xIEMMsaC2Wz2Q8TZWqebkuRhVtouNS24u
-utyaoGxBqiPLmkXMZ4B0DBByxQKBgQDTfI3OP8PoFSVsKRImX8QQEitFGadhTJEE
-7AFE2v5B8jnoa1TVSAdltbPgWWxXO79LjgJ/qQuwwSuP53mFK8dYDkVZs18PkjOe
-DHAaxBof9XP1d1+npK9gOSYyz1np2HgiWA466T4/tTNZe1BZc3nCKJCSpM13PZZz
-ZxSZO+ejCwKBgCBrgsYZCKOMOrOzN2IlXchyIxYC02aVFxu0p3lLXKT+SxSWCaOg
-PJ1MSdZBu1jV33l8GGJpQO8DB+sSPVHZG6Tm/zuIP//DaD5nEllStlOjb1g9fZ55
-3XC/pZPjeLoVdot8cYLR5ifu7HNVTZgg9iXsHOWrOlBK5RX9CDGvF0UFAoGBALYU
-PmKcM6bUzetZUO8I9FyByA8uDxWOPizmpBEMGmvciE66WcdmxDelQMCfPM+9+Oyr
-zyrp986uATtcJxuonXAKmVqpO+bEEPy56Kqiu3tYcAWeSRSkjPljOqf0LY1Ped7X
-wSinN5jSAOvBZ2QcL3AiTJZ0nRFUTPv+e5WSJdQRAoGBAIcDM/P1ktI9m6y2YuFj
-S4UGXVwjzpEzEBOXzWl7e5PBultOHqQ0yZlD66LF8ijopb4sxQjLOMpn1OqhiJr2
-jAe+S0YTdQHEvxeOZP0e6ppfiIMLbh2qWMyvPwsqFyXFcKQ9YllX0J1LKgzw1xmO
-Zefj1dkdu0Yg0mDvp7IkNx2q
------END PRIVATE KEY-----
