--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/server.crt
@@ -1,32 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFiDCCA3CgAwIBAgIUYhp72eMZaL3bPAAxL2DJ/JN1pSMwDQYJKoZIhvcNAQEL
-BQAwgYYxFzAVBgNVBAoMDkNvZGVEZXNpZ25QbHVzMR4wHAYDVQQDDBVDZXJ0aWZp
-Y2F0ZSBBdXRob3JpdHkxCzAJBgNVBAYTAkNPMRIwEAYDVQQIDAlCb2dvdMODwqEx
-EjAQBgNVBAcMCUJvZ290w4PCoTEWMBQGA1UECwwNSVQgRGVwYXJ0bWVudDAeFw0y
-MzEwMTkyMzE4MDZaFw0yNDEwMTgyMzE4MDZaMH0xFzAVBgNVBAoMDkNvZGVEZXNp
-Z25QbHVzMRUwEwYDVQQDDAxyZWRpcy1zZXJ2ZXIxCzAJBgNVBAYTAkNPMRIwEAYD
-VQQIDAlCb2dvdMODwqExEjAQBgNVBAcMCUJvZ290w4PCoTEWMBQGA1UECwwNSVQg
-RGVwYXJ0bWVudDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJpWS3Ed
-1un06XgUU+AhXvLb1cUX1GyjvxzNfmgTRQd6ix4Pr85jRIhhL3a/94fRCKac8LAt
-fBxzEGymjEMebiXeqnNEKCg4lTWUGOt5thm8BZHlDVYOG5TIx+UxAIKF+goxZNj5
-0hKOu51Yk+pUOAGJUMPjV0/tSu3xtyK9G+v7/vLq3aNoTE8q+Hleqfdc5xLBBchv
-STjpvaGxpeAZXTA2Y+yFh+K33j4elAMKi/+GR1ena1Ly7RY+FnlUEka1JzLd+QRw
-YBn+iFUoAjma54tuM3VBrhzlw59IRmyHmmg0r/xeHZkBxfY6mXiNsXt+FCFo3fDQ
-YvMBCtWOui+yHdkCAwEAAaOB9TCB8jALBgNVHQ8EBAMCBaAwEQYJYIZIAYb4QgEB
-BAQDAgZAMB0GA1UdDgQWBBQJvovzrCyo+NZh2p1MEhbLAQaotjCBsAYDVR0jBIGo
-MIGloYGMpIGJMIGGMRcwFQYDVQQKDA5Db2RlRGVzaWduUGx1czEeMBwGA1UEAwwV
-Q2VydGlmaWNhdGUgQXV0aG9yaXR5MQswCQYDVQQGEwJDTzESMBAGA1UECAwJQm9n
-b3TDg8KhMRIwEAYDVQQHDAlCb2dvdMODwqExFjAUBgNVBAsMDUlUIERlcGFydG1l
-bnSCFDMsLt43W7pMtp5gG3dCInjXHDoIMA0GCSqGSIb3DQEBCwUAA4ICAQDUJDfs
-WruiwXPsF3XUyqbJa5RSXmSIm9irF/22Xvo6m2tImXsNvRx/Xu7Wcwaxc8mok4LW
-e0k+AGXddEY20XVw7yNecsLxp9bVu4DnTPonRuoZgGBk3rAmymh96akvPryGXohi
-HWnqZ4r1NVckRj46wOPvOFPpv77wDUT0uUZZW7sBh0hd7JhMfQqcvdBiOjz4Eabk
-7nOay3fS4a+sEfJv2tH+LrPDyde5vc01FxvopmIi4HqzMMOtouDXWUp1x3Zs7okU
-GrVKR8px29TOnrzcBoMX+ftsDbFaCoHLFY5WEqpF6XSur/+tp8D3+1f9WNdAKp0c
-OpcJrLraJw6SphWIhqHFr5nldmod8Q0+F+myycWeFPOJF/PeYINMquJv7iNtItea
-3mKsYyeUkibwkhC2ozaSeMHqm6goIaV3XRbkYd4pAKmdh/MbXDA9OXymmpfMkF92
-OsFu1LPRhvxfPmp4gl1zsIj7lEq0RxI7vOmb5WbijUEsfhv+GoFiEPj0lF4KmsrE
-ghGEZ6tFTYh68El+OkPocyAk3STt7Ideh3V182Wl5mfruU2n4l20DHZ8Ow64PR8x
-Hz/zH79Fth/mg3N5xHxDrJLkGGzbdIdxYsWSPVezDTGuq29IT4nteKWAUD6RDZw0
-vXmpKqMHxYz3TXCCMpsuTow9UtqR8keut4oOnQ==
------END CERTIFICATE-----
