--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Certificates/server.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCaVktxHdbp9Ol4
-FFPgIV7y29XFF9Rso78czX5oE0UHeoseD6/OY0SIYS92v/eH0QimnPCwLXwccxBs
-poxDHm4l3qpzRCgoOJU1lBjrebYZvAWR5Q1WDhuUyMflMQCChfoKMWTY+dISjrud
-WJPqVDgBiVDD41dP7Urt8bcivRvr+/7y6t2jaExPKvh5Xqn3XOcSwQXIb0k46b2h
-saXgGV0wNmPshYfit94+HpQDCov/hkdXp2tS8u0WPhZ5VBJGtScy3fkEcGAZ/ohV
-KAI5mueLbjN1Qa4c5cOfSEZsh5poNK/8Xh2ZAcX2Opl4jbF7fhQhaN3w0GLzAQrV
-jrovsh3ZAgMBAAECggEADf91qDTEoUONSU7OrK1aJzy5Dr8yjQfs9XfN1zKoOB1t
-MLTLhTSLqrLl0ydmE/NpjXL+gXwenZXyD2iroArKgrTHUN0Fp2A6aDmeaP0I/xyG
-hMz5gKyW48+yEZ94JF5yZypDs0o1O2mmqoa2I7KvwbPql1mvOZiUCzC5I3Vzatuz
-/angoqFCUex6JCHjWrAoVYtgtRKkC9Q17SFAVyI/I1F3OoTAbtwrSKH84mELio1J
-LQ2CR6URpOiuqCxm5CK2JX1nJ/WzLJf8UTEnztekqFcRq7V54O/4OrhG2TcRF9Sb
-rOXMsnUKpXS43c9HUMZzo6x06dhe1FOK82gPrJSQAQKBgQDX9TBGI4JA3YtpUeEc
-JWZ53Pve8na6TTeP8WU4yRWzLUgjgCTF/8lV/5Y2reqpbCc3O51Y+PsKkulxLapa
-Hjf7c0pe8x2FKzOdIEd74t+DYvUEpIvy8lV89y3ipSc4cHJVCzKm8+9Xc76CVJ3w
-3otEACvnMFn3VKP2P8OZUKGmmQKBgQC29CxdxAmqJNczelpunRhoJAnK0mGY+oAk
-rDMJ19m0G7ZTM7NScQhH84nZhtQsf7JIg84mHQLa2QXvVjZV8EInmubIWdB7DKHg
-0cF45sb0hl/xtd8uUGJmlXnDJ5hx3VPy0skywfByMTLGD/JJvcbdech7Ib4Ou4xQ
-lPmChfz5QQKBgQCaAuAkhRpVTCoqOByESB/m9LQjcWM+xebEHvGDJ9Tz7W0u9CpF
-kn9xd08XO7zIirsE5tz2IMpX5iHZdymwDzD5gGUfhd4dpIuwmN6h+VBRkPcQEpFD
-95MEBfuRURrBY2AGSnAbJIgiJAdPRy1EJFHJcx3cSF1Ax/oOmbbS1fASsQKBgEZf
-rAUxWCpbW3AQu2JbCT1dx2p9/E605L+2gTdjEPnHwx1EANFQa0gJ9Nd4lnq7b9+W
-26ArfqKh4jeK6QGrljbXzbf5XElItafd51bTJmp+mb9HHx8YipE9JAYeX/REl3sK
-+5uLoCq8o/QtxBv7ZNhidOxwMURHfhStCUDcu09BAoGBAKRwdLfNNcyjsOMR8C5s
-Y3guvfXDzvsU0hskY9ttvI1i7/GpueicL/p7AQaBVgpWzdMDlkyt4VBwr37S7kU/
-1UjCZuUb9cj4vghfPs9MjMO2zGtrCieeHMFsF/tMhWbRNfCZN5IkDzUNgGSmwauo
-XYn3tPkRy08KDDNRYwkpAYOu
------END PRIVATE KEY-----
