--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Settings/redis-standalone.conf
@@ -1,58 +1,0 @@
-bind 0.0.0.0
-protected-mode no
-masterauth "12345678"
-requirepass "12345678"
-port 0
-tcp-backlog 511
-timeout 300
-tcp-keepalive 300
-daemonize no
-supervised no
-pidfile /var/run/redis.pid
-loglevel notice
-databases 16
-always-show-logo no
-stop-writes-on-bgsave-error yes
-rdbcompression yes
-rdbchecksum yes
-dbfilename dump_6380.rdb
-replica-serve-stale-data yes
-replica-read-only yes
-repl-timeout 60
-replica-priority 100
-maxclients 10000
-lazyfree-lazy-eviction no
-lazyfree-lazy-expire no
-lazyfree-lazy-server-del no
-appendonly yes
-appendfilename "appendonly.aof"
-appendfsync everysec
-no-appendfsync-on-rewrite no
-auto-aof-rewrite-percentage 100
-auto-aof-rewrite-min-size 64mb
-aof-load-truncated yes
-aof-use-rdb-preamble yes
-lua-time-limit 5000
-slowlog-log-slower-than 10000
-slowlog-max-len 128
-latency-monitor-threshold 0
-notify-keyspace-events ""
-list-max-ziplist-size -2
-activerehashing yes
-client-output-buffer-limit normal 0 0 0
-client-output-buffer-limit replica 256mb 64mb 60
-client-output-buffer-limit pubsub 32mb 8mb 60
-hz 10
-dynamic-hz yes
-rdb-save-incremental-fsync yes
-
-### TLS
-tls-port 6380
-tls-cert-file /usr/local/etc/redis/redis.crt
-tls-key-file /usr/local/etc/redis/redis.key
-tls-ca-cert-file /usr/local/etc/redis/ca.crt
-tls-auth-clients no
-tls-replication yes
-tls-cluster yes
-tls-protocols "TLSv1.2 TLSv1.3"
-tls-session-caching no
