--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/Settings/redis0.conf
@@ -1,82 +1,0 @@
-bind 0.0.0.0
-protected-mode no
-masterauth "12345678"
-requirepass "12345678"
-port 0
-tcp-backlog 511
-timeout 300
-tcp-keepalive 300
-daemonize no
-supervised no
-pidfile /var/run/redis.pid
-loglevel notice
-#logfile "/var/log/redis/redis.log"
-databases 16
-always-show-logo no
-stop-writes-on-bgsave-error yes
-rdbcompression yes
-rdbchecksum yes
-# The filename where to dump the DB
-dbfilename dump_7379.rdb
-replica-serve-stale-data yes
-replica-serve-stale-data yes
-replica-read-only yes
-repl-diskless-sync no
-repl-diskless-sync-delay 5
-repl-timeout 60
-repl-disable-tcp-nodelay no
-replica-priority 100
-maxclients 10000
-# maxmemory-policy noeviction
-lazyfree-lazy-eviction no
-lazyfree-lazy-expire no
-lazyfree-lazy-server-del no
-replica-lazy-flush no
-appendonly yes
-appendfilename "appendonly.aof"
-appendfsync everysec
-no-appendfsync-on-rewrite no
-auto-aof-rewrite-percentage 100
-auto-aof-rewrite-min-size 64mb
-aof-load-truncated yes
-aof-use-rdb-preamble yes
-lua-time-limit 5000
-cluster-config-file redis_cluster.conf
-cluster-node-timeout 15000
-cluster-enabled yes
-# cluster-replica-validity-factor 10
-# cluster-require-full-coverage yes
-slowlog-log-slower-than 10000
-slowlog-max-len 128
-latency-monitor-threshold 0
-notify-keyspace-events ""
-list-max-ziplist-size -2
-activerehashing yes
-client-output-buffer-limit normal 0 0 0
-client-output-buffer-limit replica 256mb 64mb 60
-client-output-buffer-limit pubsub 32mb 8mb 60
-hz 10
-dynamic-hz yes
-rdb-save-incremental-fsync yes
-### TLS
-tls-port 7379
-tls-cert-file /usr/local/etc/redis/redis.crt
-tls-key-file /usr/local/etc/redis/redis.key
-tls-ca-cert-file /usr/local/etc/redis/ca.crt
-tls-auth-clients yes
-
-# tls-auth-clients optional
-
-tls-replication yes
-tls-cluster yes
-
-# Explicitly specify TLS versions to support. Allowed values are case insensitive
-# and include "TLSv1", "TLSv1.1", "TLSv1.2", "TLSv1.3" (OpenSSL >= 1.1.1) or
-# any combination. To enable only TLSv1.2 and TLSv1.3, use:
-#
-tls-protocols "TLSv1.2 TLSv1.3"
-
-# By default, TLS session caching is enabled to allow faster and less expensive
-# reconnections by clients that support it. Use the following directive to disable
-# caching.
-tls-session-caching no
