--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/docker-compose.cluster.yml
@@ -1,109 +1,0 @@
-version: '3.3'
-
-networks:
-  redisnetwork:
-    driver: bridge
-
-services:
-  redis0:
-    image: redis:6.2.0
-    volumes:
-      - ./Settings/redis0.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    networks:
-      - redisnetwork
-    ports:
-      - "7379:7379"
-    restart: always
-
-  redis1:
-    image: redis:6.2.0
-    volumes:
-      - ./Settings/redis1.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    networks:
-      - redisnetwork
-    ports:
-      - "7380:7380"
-    restart: always
-
-  redis2:
-    image: redis:6.2.0
-    volumes:
-      - ./Settings/redis2.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    networks:
-      - redisnetwork
-    ports:
-      - "7381:7381"
-    restart: always
-
-  redis3:
-    image: redis:6.2.0
-    volumes:
-      - ./Settings/redis3.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    networks:
-      - redisnetwork
-    ports:
-      - "7382:7382"
-    restart: always
-
-  redis4:
-    image: redis:6.2.0
-    volumes:
-      - ./Settings/redis4.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    networks:
-      - redisnetwork
-    ports:
-      - "7383:7383"
-    restart: always
-
-  redis5:
-    image: redis:6.2.0
-    volumes:
-      - ./Settings/redis5.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    networks:
-      - redisnetwork
-    ports:
-      - "7384:7384"
-    restart: always
-
-  redis-cli:
-    image: redis:6.2.0
-    volumes:
-      - ./Scripts/start-cluster.sh:/start-cluster.sh
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    #command: redis-cli --cluster create redis0:7379 redis1:7380 redis2:7381 redis3:7382 redis4:7383 redis5:7384 --cluster-replicas 1 --tls -a 12345678 --cert /usr/local/etc/redis/redis.crt --key /usr/local/etc/redis/redis.key --cacert /usr/local/etc/redis/ca.crt
-    command: ["/bin/bash", "/start-cluster.sh"]
-    depends_on:
-      - redis0
-      - redis1
-      - redis2
-      - redis3
-      - redis4
-      - redis5
-    networks:
-      - redisnetwork
