--- packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Helpers/Server/docker-compose.standalone.yml
@@ -1,13 +1,0 @@
-version: '3'
-
-services:
-  redis-server:
-    image: redis:6.2.0
-    command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-    volumes:
-      - ./Settings/redis-standalone.conf:/usr/local/etc/redis/redis.conf
-      - ./Certificates/redis.crt:/usr/local/etc/redis/redis.crt
-      - ./Certificates/redis.key:/usr/local/etc/redis/redis.key
-      - ./Certificates/ca.crt:/usr/local/etc/redis/ca.crt
-    ports:
-      - "6380:6380"
