+++ packages/CodeDesignPlus.Net.Redis/tests/CodeDesignPlus.Net.Redis.Test/Options/InstanceTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Security.Authentication;
+using CodeDesignPlus.Net.xUnit.Helpers;
 using StackExchange.Redis;
 
 namespace CodeDesignPlus.Net.Redis.Test.Options;
