--- packages/CodeDesignPlus.Net.xUnit/src/CodeDesignPlus.Net.xUnit/Helpers/Models/AppSettings.cs
@@ -1,6 +1,0 @@
-﻿namespace CodeDesignPlus.Net.xUnit.Helpers.Models;
-
-public class AppSettings<TOptions>
-{
-    public TOptions Section { get; set; }
-}
